change command into their respective model
asd
