+++ src/main/java/seedu/address/logic/parser/ImportDrinkCommandParser.java
@@ -6,7 +6,7 @@ import static seedu.address.logic.parser.CliSyntax.PREFIX_QUANTITY;
 
 import java.util.stream.Stream;
 
-import seedu.address.logic.commands.ImportDrinkCommand;
+import seedu.address.logic.commands.stocktaker.ImportDrinkCommand;
 import seedu.address.logic.parser.exceptions.ParseException;
 import seedu.address.model.drink.Drink;
 import seedu.address.model.drink.Name;
