+++ src/main/java/seedu/address/logic/parser/SellDrinkCommandParser.java
@@ -6,7 +6,7 @@ import static seedu.address.logic.parser.CliSyntax.PREFIX_QUANTITY;
 
 import java.util.stream.Stream;
 
-import seedu.address.logic.commands.SellDrinkCommand;
+import seedu.address.logic.commands.stocktaker.SellDrinkCommand;
 import seedu.address.logic.parser.exceptions.ParseException;
 import seedu.address.model.drink.Drink;
 import seedu.address.model.drink.Name;
