+++ src/main/java/seedu/address/model/Model.java
@@ -5,8 +5,6 @@ import java.util.function.Predicate;
 import javafx.collections.ObservableList;
 import seedu.address.commons.core.LoginInfo;
 import seedu.address.model.drink.Drink;
-import seedu.address.model.drink.Price;
-import seedu.address.model.transaction.Transaction;
 import seedu.address.model.user.Password;
 import seedu.address.model.user.UserName;
 
@@ -63,28 +61,9 @@ public interface Model {
 
 
     // =============== transactions commands =====================
-    /**
-     * Decreases the quantity of the drink in the {@code transaction}.
-     */
-    void sellDrink(Transaction transaction);
-
-    /**
-     * Increases the quantity of the dirnk in the {@code transaction}
-     */
-    void importDrink(Transaction transaction);
-
-    /**
-     * Returns an unmodifiable view of the transaction list
-     */
-    ObservableList<Transaction> getTransactionList();
-    String getTransactions(); //TODO: will be removed when UI is up
 
     // ================ analysis commands ==========================
 
-    /**
-     * Returns the total cost incurred.
-     */
-    Price analyseCosts();
