+++ src/main/java/seedu/address/model/user/accountant/AccountantModel.java
@@ -1,9 +1,15 @@
 package seedu.address.model.user.accountant;
 
 import seedu.address.model.Model;
+import seedu.address.model.drink.Price;
 
 /**
  * This is the API manager for accountant role
  */
 public interface AccountantModel extends Model {
+    /**
+     * Returns the total cost incurred.
+     */
+    Price analyseCosts();
+
 }
