+++ src/main/java/seedu/address/model/user/admin/AdminModelManager.java
@@ -1,9 +1,15 @@
 package seedu.address.model.user.admin;
 
+import java.util.List;
+
+import javafx.collections.FXCollections;
+import javafx.collections.ObservableList;
 import seedu.address.model.LoginInfoManager;
 import seedu.address.model.ModelManager;
 import seedu.address.model.ReadOnlyInventoryList;
 import seedu.address.model.UserPrefs;
+import seedu.address.model.drink.Price;
+import seedu.address.model.transaction.Transaction;
 import seedu.address.model.transaction.TransactionList;
 import seedu.address.model.user.AuthenticationLevel;
 import seedu.address.model.user.Password;
@@ -23,6 +29,54 @@ public class AdminModelManager extends ModelManager implements AdminModel {
         return false;
     }
 
+
+
+    //=====================Stock taker commands====================
+    @Override
+    public void sellDrink(Transaction transaction) {
+        Price defaultSalePrice = inventoryList.getDefaultSellingPrice(transaction.getDrinkTransacted());
+
+        Price defaultAmountTransacted = new Price(Float.toString(defaultSalePrice.getValue()
+                * transaction.getQuantityTransacted().getValue()));
+        transaction.setAmountMoney(defaultAmountTransacted);
+        recordTransaction(transaction);
+
+        inventoryList.decreaseQuantity(transaction.getDrinkTransacted(), transaction.getQuantityTransacted());
+    }
+
+    @Override
+    public void importDrink(Transaction transaction) {
+        Price defaultCostPrice = inventoryList.getDefaultCostPrice(transaction.getDrinkTransacted());
+
+        Price defaultAmountTransacted = new Price(Float.toString(defaultCostPrice.getValue()
+                * transaction.getQuantityTransacted().getValue()));
+        transaction.setAmountMoney(defaultAmountTransacted);
+        recordTransaction(transaction);
+
+        inventoryList.increaseQuantity(transaction.getDrinkTransacted(), transaction.getQuantityTransacted());
+    }
+
+    private void recordTransaction(Transaction transaction) {
+        transactionList.addTransaction(transaction);
+    }
+
+
+    /**
+     * Returns an unmodifiable view of the list of {@code Transaction} backed by the internal list of
+     * {@code transactionList}
+     */
+    @Override
+    public ObservableList<Transaction> getTransactionList() {
+        List<Transaction> transactions = transactionList.getTransactions();
+        return FXCollections.unmodifiableObservableList(FXCollections.observableList(transactions));
+    }
+
+    @Override
+    public String getTransactions() {
+        return transactionList.toString();
+    }
+
+    //=====================Manager command=========================
     @Override
     public void createNewAccount (UserName userName, Password password, AuthenticationLevel authenticationLevel) {
         loginInfoManager.createNewAccount (userName, password, authenticationLevel);
@@ -32,4 +86,11 @@ public class AdminModelManager extends ModelManager implements AdminModel {
     public void deleteAccount (UserName userName) {
         loginInfoManager.deleteAccount (userName);
     }
+
+    //=====================Accountant command======================
+    @Override
+    public Price analyseCosts() {
+        return analysis.analyseCost();
+    }
+
 }
